new training plan modal and refactoring
